Everything works up to creating an invoice.
Right now the app can pair with BitPay using a pairing code. The
"Bitcoin with Bitpay" tender type appears in the register and calls the
example activity. Things to be done from this point:

Cleaning up the app so that the Tender is not created by hand, improving
the user interface, creating invoices, handling IPNs.